feat: Update US tax brackets and rules for 2025
Includes changes to tax brackets, standard deductions, and wage base for the United States to reflect 2025 projections.

Introduces support for married filing status by adding `exemptAmountMarried` and `bracketsMarried` to `TaxBracket` and `CountryRules` types.

Refactors surcharge logic to be more generic and applicable to taxes like Soli and Church Tax, with an added threshold for conditional application.